Add files via upload
Attached is the dashboard prepared on Power BI. It is the Analysis for ITV Hotels chain. The Management wants to check the sales for last 3 months. It attracts the customers from 7 different ways. The Spread of the business is across 8 Metro cities inclusive of multiple lavish properties and 4 class of the rooms. The dashboard is dynamic. Please let me know what you think in the comments.