Add files via upload
This is a EDA performed on Wrist Watches data I got on Kaggle.com. I found this topic and data a little bit interesting as I have some interest in Watches. 
This data required a lot of cleaning which I did on Excel as it's a simpler way. Several columns like 'Year of production', 'Condition', 'Price' had null values or blank cells hence replaced with the relevant data as per my understanding. Also columns like 'Fast Shipper', 'Trusted Seller' and 'Punctuality' were having binary values so I replaced them with No for 0 and Yes for 1. 
Once data uploaded on Jupyter, I added a column named 'Revenue in Millions' which calculated as per Price and Quantity Sold.

Problem Statement: 
The Agency had approximately 25% stock unsold which was affecting the Revenue and hence they were unable to check the Profits(however they have not provided us with BuyPrice though)
Conclusion basis of the EDA:
There are some type of watches which People are ready to buy at higher price but Agency has priced them less hence they People might be considering as Cheap as per their Standard and avoiding them.

Following are the questions helped in conclusion
 
#Top_10_Most_Valued_Watches
#DialShapewise_Sale
#Conditionwise_Qty_sold
#Genderwise_sale
#DialColorwise_sale
#BraceletMaterialwise_sale
#Sale by Qty Sold
#Total No of Brands